backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
cryptography>=3.4.7
requests>=2.26.0
python-dotenv>=0.19.0 